Encapsulate block plugin ID state getter/setter and getNextPluginId() logic in a farm_setup.wizard service.
