+++ modules/core/setup/src/Form/FarmSetupBlockForm.php
@@ -4,11 +4,7 @@ declare(strict_types=1);
 
 namespace Drupal\farm_setup\Form;
 
-use Drupal\Core\DependencyInjection\AutowireTrait;
 use Drupal\Core\Form\FormStateInterface;
-use Drupal\Core\State\StateInterface;
-use Drupal\farm_setup\SetupFormPluginManager;
-use Symfony\Component\DependencyInjection\Attribute\Autowire;
 
 /**
  * Setup block form.
@@ -16,16 +12,6 @@ use Symfony\Component\DependencyInjection\Attribute\Autowire;
  * @ingroup farm
  */
 class FarmSetupBlockForm extends FarmSetupForm {
-
-  use AutowireTrait;
-
-  public function __construct(
-    #[Autowire(service: 'plugin.manager.setup_form')]
-    SetupFormPluginManager $setupFormPluginManager,
-    protected StateInterface $state,
-  ) {
-    parent::__construct($setupFormPluginManager);
-  }
 
   /**
    * {@inheritdoc}
@@ -42,9 +28,8 @@ class FarmSetupBlockForm extends FarmSetupForm {
     // This performs the same logic as the parent method, but does not perform
     // any redirections. Instead, it saves the next plugin to state so that it
     // is shown next time the block is loaded.
-    $plugin_id = $form_state->getValue('plugin_id');
-    $next_plugin_id = $this->getNextPluginId($plugin_id);
-    $this->state->set('farm_setup.block', $next_plugin_id);
+    $next_plugin_id = $this->setupWizard->getNextPluginId($form_state->getValue('plugin_id'));
+    $this->setupWizard->setBlockPluginId($next_plugin_id);
     if (is_null($next_plugin_id)) {
       $this->completeMessage();
     }
